seperated route configuration into new package
