--- restaurants.go
@@ -1,9 +1,0 @@
-package main
-
-// GetNames '/names' returns all restuarant names
-// func GetNames(w http.ResponseWriter, req *http.Request) {
-
-// 	var rest models.Restaurants
-// 	db.Find(&rest)
-// 	json.NewEncoder(w).Encode(&rest)
-// }
